Add a per-file ratchet for expect("int cast") sites and clear json5, image codecs and elf
test/internal/source-lints/int-cast-expects.test.ts pins the number of
`.expect("int cast")` sites per Rust file (1151 sites in 273 files after
this commit) so new ones fail the source lints and removals have to lower
the limits file.

json5: reject documents longer than i32::MAX up front (DocumentTooLarge,
like the JSON parser) and build token locations with usize2loc.

image codecs: a decoder-reported c_int dimension converts through
positive_dimension (DecodeFailed), kernel dimensions through
kernel_dimension (TooManyPixels), and rotate takes the u16 degrees the
pipeline already stores.

elf: every offset read from a --compile-executable-path template goes
through file_range, which bounds-checks it against the file and returns
InvalidElfFile. The .bun sh_offset, e_phoff, .shstrtab offset and
e_shstrndx were previously used unchecked, and several header sums could
overflow; a corrupt template used to abort the build with a slice-index
panic.

+++ src/runtime/image/Image.rs
@@ -1945,7 +1945,7 @@ impl PipelineTask {
     fn apply_pipeline(&self, d: &mut codecs::Decoded) -> Result<(), codecs::Error> {
         let p = &self.pipeline;
         if p.rotate != 0 {
-            let next = codecs::rotate(&d.rgba, d.width, d.height, u32::from(p.rotate))?;
+            let next = codecs::rotate(&d.rgba, d.width, d.height, p.rotate)?;
             // Assignment drops
             // the old `Vec<u8>`/owned buffer.
             d.rgba = next.rgba;
@@ -2071,7 +2071,7 @@ fn apply_orientation(
     if t.rotate != 0 {
         // Swap pixel slots only — `next` carries no ICC profile, and the
         // one on `d` (set by decode) must survive EXIF auto-orient.
-        let next = codecs::rotate(&d.rgba, d.width, d.height, u32::from(t.rotate))?;
+        let next = codecs::rotate(&d.rgba, d.width, d.height, t.rotate)?;
         d.rgba = next.rgba;
         d.width = next.width;
         d.height = next.height;
